Update layers.py: remove a mypy: ignore
`mypy: ignore` is not (yet!) actually a feature of the mypy typechecker. However, once I implemented that feature in [my pull-request](https://github.com/python/mypy/pull/17875), it was reported to me that the `type: ignore` you get here is actually unused anyway! (I guess that makes sense, or yous would have noticed the type error it was supposed to be suppressing get raised anyway.) So, it can just be removed.